Kills to raid 1-2 player clan = 28~ Kills to raid 3-4 player clan = 25 Kills to raid 5-6 player clan = 22~ Kills to raid 7-8 player clan = 20

+++ clans/src/main/java/me/mykindos/betterpvp/clans/clans/menus/ClanMenu.java
@@ -169,7 +169,7 @@ public class ClanMenu extends Menu {
 
     private List<ClanEnemy> getTopEnemiesByDominance() {
         List<ClanEnemy> enemies = clan.getEnemies();
-        enemies.sort(Comparator.comparingInt((ClanEnemy e) -> Math.abs(e.getDominance())).reversed());
+        enemies.sort(Comparator.comparingDouble((ClanEnemy e) -> Math.abs(e.getDominance())).reversed());
 
         return enemies.subList(0, Math.min(5, enemies.size()));
     }
